+++ Editor/Scripts/Typed/BoolStringGameEventInspector.cs
@@ -1,0 +1,8 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(BoolStringGameEvent))]
+  public class BoolStringGameEventInspector : TwoTypeGameEventInspector<bool, string>
+  { }
+}