+++ Editor/Scripts/Typed/StringStringStringGameEventInspector.cs
@@ -1,0 +1,8 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(StringStringStringGameEvent))]
+  public class StringStringStringGameEventInspector : ThreeTypeGameEventInspector<string, string, string>
+  { }
+}