+++ Runtime/Scripts/TypedGameEvent/Typed/BoolStringGameEvent.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/Bool-String Game Event")]
+  public class BoolStringGameEvent : TwoTypeGameEvent<bool, string>
+  { }
+}