+++ Runtime/Scripts/TypedGameEvent/Typed/StringStringStringGameEvent.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/String-String-String Game Event")]
+  public class StringStringStringGameEvent : ThreeTypeGameEvent<string, string, string>
+  { }
+}